fix golang lint error

+++ server/storage/objectstorage/reader_impl.go
@@ -623,11 +623,11 @@ func (f *MinioFileReaderAdv) readDataBlocksUnsafe(ctx context.Context, opt stora
 		// 2. We have collected some data and no more blocks available
 		// 3. We reached the end of available blocks
 		// 4. We encounter expected read error, data incomplete verification error / network/service unavailable, etc.
-		if totalCollectedSize >= int64(maxBytes) {
+		if totalCollectedSize >= maxBytes {
 			logger.Ctx(ctx).Debug("reached max collected size limit",
 				zap.String("segmentFileKey", f.segmentFileKey),
 				zap.Int64("totalCollectedSize", totalCollectedSize),
-				zap.Int64("maxBytes", int64(maxBytes)))
+				zap.Int64("maxBytes", maxBytes))
 			break
 		}
